pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conformal-econ"
version = "0.1.0"
description = "Conformal prediction intervals for economic time series forecasting"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Adriannie" }]
dependencies = [
    "numpy>=1.26,<2.0",
    "pandas>=2.1,<3.0",
    "statsmodels>=0.14,<1.0",
    "scikit-learn>=1.4,<2.0",
    "xgboost>=2.0,<3.0",
    "torch>=2.2,<3.0",
    "MAPIE>=0.8,<1.0",
    "transformers>=4.40,<5.0",
    "streamlit>=1.32,<2.0",
    "plotly>=5.20,<6.0",
    "fredapi>=0.5,<1.0",
    "python-dotenv>=1.0,<2.0",
    "pyarrow>=15.0,<16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "mypy>=1.9,<2.0",
    "ruff>=0.4,<1.0",
    "pre-commit>=3.7,<4.0",
    "pandas-stubs>=2.1,<3.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=conformal_econ --cov-report=term-missing"

[tool.coverage.run]
source = ["conformal_econ"]
omit = ["*/tests/*"]
